Move legacy Porytiles code into a legacy directory
