+++ Porytiles/include/porytiles/legacy/emitter.h
